+++ bin/fast.dart
@@ -1,6 +1,7 @@
 import 'package:args/command_runner.dart';
 import 'package:flutter_fast_cli/src/commands/build/build.dart';
 import 'package:flutter_fast_cli/src/commands/commands.dart';
+import 'package:flutter_fast_cli/src/commands/wizard/wizard.dart';
 
 void main(List<String> arguments) {
   CommandRunner(
@@ -8,6 +9,7 @@ void main(List<String> arguments) {
     "CLI for building Flutter apps fast",
   )
     ..addCommand(CreateApp())
+    ..addCommand(Wizard())
     ..addCommand(Clean())
     ..addCommand(Build())
     ..run(arguments);